+++ src/forkAndExecuteSpell.ts
