+++ src/periphery/reporter/ReportSender.ts
@@ -3,24 +3,26 @@ import { HttpClient } from '@sparkdotfi/common-universal/http-client'
 import { Logger } from '@sparkdotfi/common-universal/logger'
 
 export class ReportSender {
+  reporters: IReporter[]
+
   constructor(
     private readonly wehbookUrl: string | undefined,
     private readonly logger: Logger,
     private readonly httpClient: HttpClient,
   ) {
     this.logger = logger.for(this)
+    this.reporters = this.createReporters()
   }
 
   async send(reports: Report[]): Promise<void> {
     for (const report of reports) {
-      const reporters = this.getReporters()
-      for (const reporter of reporters) {
+      for (const reporter of this.reporters) {
         await reporter.report(report)
       }
     }
   }
 
-  private getReporters(): IReporter[] {
+  private createReporters(): IReporter[] {
     const consoleReporter = new ConsoleReporter(this.logger, 'info')
 
     if (!this.wehbookUrl) {
